+++ app/src/main/java/com/example/cpbyte_portal/presentation/ui/theme/Color.kt
@@ -9,3 +9,16 @@ val Pink80 = Color(0xFFEFB8C8)
 val Purple40 = Color(0xFF6650a4)
 val PurpleGrey40 = Color(0xFF625b71)
 val Pink40 = Color(0xFF7D5260)
+
+val textPrimaryColor = Color(0xffE0F2FE)
+val textSecondaryColor = Color(0xff93C5FD)
+
+val inputFieldBgColor = Color(0xff1E293B)
+
+val fieldTextPrimaryColor = Color(0xffE2E8F0)
+val fieldTextSecondaryColor = Color(0xff7DD3FC)
+
+val cardBgColor = Color(0xff0F172A)
+val cardBorderColor = Color(0xff3B82F6)
+
+val imageTintColor = Color(0x403B82F6)
